Modificación de consultas de estadisticas

+++ ATMScoreBoard/ATMScoreBoard.Shared/DTOs/EstadisticaEquipoColRanking.cs
@@ -9,7 +9,7 @@ namespace ATMScoreBoard.Shared.DTOs
     public class EstadisticaEquipoColRanking
     {
         public int Id { get; set; }
-        public long PosicionRanking { get; set; }
+        public int PosicionRanking { get; set; }
         public string? JugadorA { get; set; }
         public string? JugadorB { get; set; }
         public int PuntosRanking { get; set; }
